fix fastFileComparison for network files
BUG: 403089
FIXED-IN:1.7.90

This fixes the handling of tempfiles created during network comparsion.
Previously these would be accessed by name alone causing an error.
Now QTemparyFile is used were needed.
Direct use of QFile for file operations is now depriated as this is
diffcualt to get right given the use of a tempFile in some cases.
*open,close, read routines created in FileAccess to currectly handle the
tempfile logic and error messages.
*Fix this issue exists within FileAccess::readFile() as well.