Dont use QUrl::toDisplay string in contex menu.
This is being done for the same reasons as for the main app.
Although the menu plugin is generally only expected to run on Linux Qt sometimes decides to leave file names encoded.
This issue affects users with non-latin characters in the file name.
The errant encoding is often longer then the orignal name.
Resulting in over sized context menus.
The underling cause of the context menu issue lies in the Qt/Frameworks code that displays it.
There seems to be no consept of shorting displayed strings when they are very long.
There is only so much kdiff3 can do regarding this issue.