Don't add raw pointers to a std::list they won't be deleted.
With then convertion to boost::signal2 mOptionItemList is now sololy a way to auto delete option items.
The QObject hybrids in OptionDialog need no longer be added to this list.
This in turn removes Qt imposed restrictions on copying and moving that forced the use of raw pointers.
Altough non-leaky in the current implementation using raw pointers with std::list is bad practice.