Move mergeOneLine to MergeLine
mergeOneLine only needs read access to Diff3Line but requires write access to MergeLine.
Therefor it belongs in MergeLine as it deals with internal details of that class.

Expose new public API for Diff3Line we can check if it has fine diff data without direct access to its memmbers.