Move DirectoryInfo to global scope.
This makes the code cleaner and less confusing we only need one DirectoryInfo.
Keep shared pointer status for now maybe reviewed later but is out of scope.
KDiff3 has multiple code paths for starting a compare so the old intialization did not cover all cases.
Move this first time setup to the default iniialization.
This is safe for most code. Constuctors for other global/static variables however can not depend on this having taken place.

BUG:439676
FIXED-IN:1.9.3