lint: cppcheck fixes
cppcheck is not run routinely on this project due to lack of proper compile_commands.json support.
False parsing errors are common when attempting to run it. This noise masks any real messages there maybe.
Further a number of checks will not pass cleanly do to design/stylistic choices.
The suppression file it this commit only blocks the main offenders.