Block named pipes
This prevents KIO from recieving files names with pipe protocol.
This protocol is not supportted.
Qt5 misbehaves in intresting ways when trying use QFile with a pipe so we can't easiily support this.
Additionallly kdiff3 itself assumes it is reading from a file-like stream that can be re-opened after closing and has a definite size.