Use QThread instead of QRunnable
QRunnable is not enough here be may appear to work depending on configuration.
That is an artifact of the threads being short-lived. Never the less such code is
not safe because of a potential race condition on exit.

The use of Qt's parent system establishes order of destruction for QObjects.
Allowing us to guarantee clearance of these helper threads and their accompanying
objects before the associated DiffTextWindow begins teardown.