Fix starting with command line parameters
With the former code there was always an error shown on startup even if
the files could be opened and read. This because the the signal
'wasCancelled' gets set when the function 'ProgressDialog::cancel()'
gets called.

I'm not 100% sure, if this doesn't introduce some other problem. But I tested it with file and folder diff from command line and it seems to work fine.

I'm open for suggestions for improvement. ;o)