Use utf-16 wide char literals
Functionally this is a no-op since we are dealing with code points thet align for all ASCII compatible codecs.
UTF-8, Latin1 or ASCII is used in practice by gcc, clang and MSVC when encoding them.
Source files are UTF-8.

These values will now be 16 bit and a ganerteed by c++11 to be in UTF-16 encoding.